requirements.txt
Flask==3.0.3
Flask-SQLALchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-Limiter==3.7.0
Flask-Migrate==4.0.7
PyMySQL==1.0.1
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow>=12.1.1
stripe==14.4.1
paypalrestsdk==1.13.3